Added comments and classes separation
